Initial migration of device communication code to separate class (apart from ReadPane UI class) is complete. No support for multiple cards or theft bit.
