+++ nbproject/private/private.xml
@@ -4,10 +4,9 @@
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
             <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/ProjectNotes.txt</file>
-            <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/AppConstants.java</file>
+            <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/SkyRfidJavaApp.java</file>
             <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/RfidNativeInterface.java</file>
             <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/TagReader.java</file>
-            <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/TheftBitWorker.java</file>
             <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/TagEncoding.java</file>
             <file>file:/C:/Users/MichalG/Documents/NetBeansProjects/SkyRfidJavaApp/src/skyrfidjavaapp/ReadPane.java</file>
         </group>
