+++ src/skyrfidjavaapp/DeviceErrorCodes.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2017 Michal G. <Michal.G at cogitatummagnumtelae.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package skyrfidjavaapp;
+
+/**
+ *
+ * @author Michal G. <Michal.G at cogitatummagnumtelae.com>
+ */
+public class DeviceErrorCodes {
+    public static final String ERR_NO_HANDLE = "cannot find the USB device";
+    public static final String ERR_MULTIPLE_CARDS = "multiple cards detected in single-read mode";
+}
