+++ src/skyrfidjavaapp/RfidNativeInterface.java
@@ -30,19 +30,43 @@ public interface RfidNativeInterface extends Library {
     
     
     /**
-     * Initialize the RFID device
-     * @param port
-     * @param baud_rate
-     * @return handle to device
+     * Open the port to the RFID device
+     * @param port This application is designed exclusively for the USB reader.
+     * See manufacturer's documentation for serial device.
+     * @param baud_rate Ignored for USB.
+     * @return handle to device. >0 = success. <0 = fail.
      */
     int fw_init(int port, long baud_rate);   
     
-    
+    /**
+     * Configures the USB device to read the desired type of tag.
+     * @param handle handle to USB device
+     * @param card_type  This application uses ISO15693 exclusively.
+     * See manufacturer's documentation for ISO14443 Type A and B.
+     * @return 0 = success. Nonzero = fail.
+     */
     int fw_config_card(int handle,char card_type); 
     
+    /**
+     * 
+     * @param handle handle to USB device
+     * @param single_multi_flag
+     * @param app_id Can be 0
+     * @param mask_length
+     * @param returned_length
+     * @param returned_buffer
+     * @return 
+     */
     int fw_inventory(int handle, char single_multi_flag, char app_id, 
             char mask_length, char[] returned_length, char[] returned_buffer);   
     
+    /**
+     * 
+     * @param handle handle to USB device
+     * @param flags
+     * @param card_id
+     * @return 
+     */
     int fw_select_uid(int handle,char flags, char[] card_id); 
     
     
@@ -60,7 +84,7 @@ public interface RfidNativeInterface extends Library {
      * @param card_id
      * @param returned_data_length
      * @param returned_data
-     * @return 
+     * @return 0 = success. Nonzero = fail.
      */
     int fw_readblock(int handle,char flags, char startblock,
             char number_of_blocks_to_read, char[] card_id, char[] returned_data_length, char[] returned_data);
@@ -70,7 +94,7 @@ public interface RfidNativeInterface extends Library {
      * @param handle handle to the RFID device
      * @param flags
      * @param card_id
-     * @return 
+     * @return 0 = success. Nonzero = fail.
      */
     int fw_stay_quiet(int handle, char flags, char[] card_id);
 //    st=fw_stay_quiet(icdev,0x22,&UID[1])
